Improve debugging support
When this error is encountered in the field, I see it via honeybadger, but I have none of the context about what unit it was called in, and what unit it was being converted to. Help the debugging by adding some info to the exception. I'm not entirely sure that 'toString' is what I want here - perhaps others in the community can help make that the right expression?