Update subset for sparse matrix
The method for replacing a subset in a sparse matrix is changed in order to take into account the indexes used as arguments, solving an issue where non-consecutive indexes would give a different result compared to the one given when using a dense matrix.

Unit tests are added for this issue and 2 other tests are changed in order to correctly display the result of this operation.