Add range information to SymbolNode and ConstantNode
The parser knows the position of the nodes in
the original text and can as well return this
information in the node objects.

This is useful when implementing the MathJS syntax
highlighting in a rich text editor.

One drawback of adding "range" property to the node
object is that two semantically equal nodes might not
be equal when parsed from different strings as their
"range" properties might differ. This caused some
failures of `assert.deepStrictEqual` in tests.