Fix: don't throw exception when filter/map/forEach an empty array (fixes: #3564)
In #3564 some examples were given where `evaluate("map([], f(x, b) = x)")` or similar would throw an exception instead of returning an empty array.

The root cause was an out-of-bounds access attempt in `optimizeCallback()` when trying to get the first value in the array/matrix; most likely introduced in #3256 or one of the later updates to that code.

This is resolved by first checking if the array or matrix is empty, and if so immediately returning a default response (which will never be used).

## Additional details on impact of bug

As this is in a shared function, this also impacts `filter` and `forEach` as well as `map` as initially reported.

As this was only in the path optimising non-unary typed function, then only specific callback forms would trigger the bug:

- `evaluate('map([], x + 1)')` would NOT trigger the issue as that gives an untyped function
- `evaluate('map([], f(x) = x +  1)')` WOULD trigger the issue as the function is now typed (and non-unary)

## Testing

Additional unit tests were added to verify the cases in question (and the cases that already worked).

Verified that:

- the new unit tests for typed callbacks failed before the fix
- the new unit tests for untyped callbacks passed before the fix (for reasons described above)
- both typed and untyped callbacks pass after the fix
- all other unit tests - using `npm run test:all` continue to run and pass

###  New Test results (with spec reporter)

```
empty arrays and matrices
      with untyped callbacks
        ✔ should return an empty array when filtering an empty array
        ✔ should return an empty array when mapping an empty array
        ✔ should not throw forEaching an empty array
      with typed callbacks
        ✔ should return an empty array when filtering an empty array
        ✔ should return an empty array when mapping an empty array
        ✔ should not throw forEaching an empty array
```

## Performance

The change in #3256 was created to improve performance, so it may be worth reviewing this change with that in mind.  I believe it should have minimal impact as it only adds 1 array access and 1 `if` (and perhaps 1 variable assignment that should be optimised out).

That said, I have not run the benchmarks or compared the results as part of this PR.